Move noise functions and tests from preprocess to new filter subpackage
