+++ echopype/__init__.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 from _echopype_version import version as __version__  # noqa
 
-from . import calibrate, consolidate, mask, preprocess, utils
+from . import calibrate, consolidate, filter, mask, preprocess, utils
 from .convert.api import open_raw
 from .echodata.api import open_converted
 from .echodata.combine import combine_echodata
@@ -20,6 +20,7 @@ __all__ = [
     "combine_echodata",
     "calibrate",
     "consolidate",
+    "filter",
     "mask",
     "preprocess",
     "utils",
