+++ echopype/filter/__init__.py
@@ -1,0 +1,5 @@
+from .api import remove_noise
+
+__all__ = [
+    "remove_noise",
+]
