+++ echopype/preprocess/__init__.py
@@ -1,7 +1,6 @@
-from .api import compute_MVBS, compute_MVBS_index_binning, remove_noise
+from .api import compute_MVBS, compute_MVBS_index_binning
 
 __all__ = [
     "compute_MVBS",
     "compute_MVBS_index_binning",
-    "remove_noise",
 ]
