+++ echopype/preprocess/api.py
@@ -8,7 +8,6 @@ import xarray as xr
 
 from ..utils.prov import echopype_prov_attrs
 from .mvbs import get_MVBS_along_channels
-from .noise_est import NoiseEst
 
 
 def _set_MVBS_attrs(ds):
@@ -212,72 +211,6 @@ def compute_MVBS_index_binning(ds_Sv, range_sample_num=100, ping_num=100):
     ds_MVBS["frequency_nominal"] = ds_Sv["frequency_nominal"]  # re-attach frequency_nominal
 
     return ds_MVBS
-
-
-def estimate_noise(ds_Sv, ping_num, range_sample_num, noise_max=None):
-    """
-    Remove noise by using estimates of background noise
-    from mean calibrated power of a collection of pings.
-
-    See ``remove_noise`` for reference.
-
-    Parameters
-    ----------
-    ds_Sv : xr.Dataset
-        dataset containing ``Sv`` and ``echo_range`` [m]
-    ping_num : int
-        number of pings to obtain noise estimates
-    range_sample_num : int
-        number of samples along the ``range_sample`` dimension to obtain noise estimates
-    noise_max : float
-        the upper limit for background noise expected under the operating conditions
-
-    Returns
-    -------
-    A DataArray containing noise estimated from the input ``ds_Sv``
-    """
-    noise_obj = NoiseEst(ds_Sv=ds_Sv.copy(), ping_num=ping_num, range_sample_num=range_sample_num)
-    noise_obj.estimate_noise(noise_max=noise_max)
-    return noise_obj.Sv_noise
-
-
-def remove_noise(ds_Sv, ping_num, range_sample_num, noise_max=None, SNR_threshold=3):
-    """
-    Remove noise by using estimates of background noise
-    from mean calibrated power of a collection of pings.
-
-    Reference: De Robertis & Higginbottom. 2007.
-    A post-processing technique to estimate the signal-to-noise ratio
-    and remove echosounder background noise.
-    ICES Journal of Marine Sciences 64(6): 1282–1291.
-
-    Parameters
-    ----------
-    ds_Sv : xr.Dataset
-        dataset containing ``Sv`` and ``echo_range`` [m]
-    ping_num : int
-        number of pings to obtain noise estimates
-    range_sample_num : int
-        number of samples along the ``range_sample`` dimension to obtain noise estimates
-    noise_max : float
-        the upper limit for background noise expected under the operating conditions
-    SNR_threshold : float
-        acceptable signal-to-noise ratio, default to 3 dB
-
-    Returns
-    -------
-    The input dataset with additional variables, including
-    the corrected Sv (``Sv_corrected``) and the noise estimates (``Sv_noise``)
-    """
-    noise_obj = NoiseEst(ds_Sv=ds_Sv.copy(), ping_num=ping_num, range_sample_num=range_sample_num)
-    noise_obj.remove_noise(noise_max=noise_max, SNR_threshold=SNR_threshold)
-    ds_Sv = noise_obj.ds_Sv
-
-    prov_dict = echopype_prov_attrs(process_type="processing")
-    prov_dict["processing_function"] = "preprocess.remove_noise"
-    ds_Sv = ds_Sv.assign_attrs(prov_dict)
-
-    return ds_Sv
 
 
 def regrid():
